♻️ Move InlineSaveIndicator and useInlineSave to shared locations
